+++ app/mypage/page.tsx
@@ -1,8 +1,22 @@
+import { redirect } from "next/navigation";
 import { createClient } from "../../lib/utils/server";
 
-export default async function Mypage() {
-    const supabase = await createClient();
-    const { data: instruments } = await supabase.from("instruments").select();
+export function generateMetadata() {
+    return {
+        title: "마이페이지 - 축제가자",
+    };
+}
 
-    return <pre>{JSON.stringify(instruments, null, 2)}</pre>;
+export default async function Mypage() {
+    const supabase = createClient();
+    const {
+        data: { user },
+    } = await (await supabase).auth.getUser();
+
+    // 로그인 X 라면 로그인 페이지로 리다이렉션
+    if (!user) {
+        redirect("/login");
+    }
+
+    return <div>마이페이지</div>;
 }
